[server] Move BlockedUserFilter out of ee
